+++ src/main/kotlin/com/lukastomoszek/idea/codemetricsvisualization/context/PsiUtils.kt
@@ -19,27 +19,28 @@ package com.lukastomoszek.idea.codemetricsvisualization.context
 import com.intellij.openapi.application.readAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
-import com.intellij.psi.PsiDocumentManager
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiFile
-import com.intellij.psi.PsiMethod
+import com.intellij.psi.*
 import com.intellij.psi.util.PsiTreeUtil
 
 object PsiUtils {
 
     suspend fun getContainingMethodFqn(element: PsiElement): String? {
         return readAction {
-            val method = PsiTreeUtil.getParentOfType(element, PsiMethod::class.java, false) ?: return@readAction null
-            val containingClass = method.containingClass ?: return@readAction null
+            var currentElement: PsiElement? = element
+            while (currentElement != null) {
+                val method = PsiTreeUtil.getParentOfType(currentElement, PsiMethod::class.java, false)
+                if (method == null) return@readAction null
 
-            val methodName = method.name
-            val qualifiedClassName = containingClass.qualifiedName
+                val containingClass = method.containingClass
+                if (containingClass != null && containingClass !is PsiAnonymousClass) {
+                    val methodName = method.name
+                    val qualifiedClassName = containingClass.qualifiedName
+                    return@readAction if (qualifiedClassName != null) "$qualifiedClassName.$methodName" else methodName
+                }
 
-            if (qualifiedClassName != null) {
-                "$qualifiedClassName.$methodName"
-            } else {
-                methodName
+                currentElement = method.parent
             }
+            return@readAction null
         }
     }